pyproject.toml
[project]
name = "drf-2fa-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.4",
    "djangorestframework-simplejwt>=5.3.1",
    "djangorestframework>=3.15.2",
    "loguru>=0.7.3",
    "python-dotenv>=1.0.1",
    "pyjwt>=2.10.1",
    "redis>=5.2.1",
    "psycopg[binary]>=3.2.3",
    "gunicorn>=23.0.0",
    "drf-spectacular>=0.28.0",
]

[dependency-groups]
dev = [
    "pytest-django>=4.9.0",
    "pytest>=8.3.4",
    "freezegun>=1.5.1",
]
